#28 - feat : [BE] LoadTask 엔티티 수정, ReadyStatusEntity 삭제

+++ backend/src/main/java/org/ssafy/load/dao/LoadTaskRepository.java
@@ -4,6 +4,5 @@ package org.ssafy.load.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.ssafy.load.domain.LoadTaskEntity;
 
-public interface LoadTaskRepository extends JpaRepository<LoadTaskEntity, Long> {
-
+public interface LoadTaskRepository extends JpaRepository<LoadTaskEntity, Integer> {
 }
